+++ packages/dependency-extraction-webpack-plugin/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### New Features
+
+-   Add `@wordpress/boot`,  `@wordpress/theme`, and `@wordpress/route` to list of bundled packages. These won't be externalized to WP/Gutenberg by the plugin and are instead included in the build output.
+
 ## 6.39.0 (2026-01-29)
 
 ## 6.38.0 (2026-01-16)